+++ packages/flutter/lib/src/integrations/flutter_error_integration.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/foundation.dart';
 import 'package:sentry/sentry.dart';
 import '../sentry_flutter_options.dart';
@@ -22,7 +24,7 @@ class FlutterErrorIntegration implements Integration<SentryFlutterOptions> {
   @override
   void call(Hub hub, SentryFlutterOptions options) {
     _defaultOnError = FlutterError.onError;
-    _integrationOnError = (FlutterErrorDetails errorDetails) async {
+    _integrationOnError = (FlutterErrorDetails errorDetails) {
       final exception = errorDetails.exception;
 
       options.log(
@@ -88,7 +90,7 @@ class FlutterErrorIntegration implements Integration<SentryFlutterOptions> {
           stackTrace = getCurrentStackTrace();
           hint.addAll({TypeCheckHint.currentStackTrace: true});
         }
-        await hub.captureEvent(event, stackTrace: stackTrace, hint: hint);
+        unawaited(hub.captureEvent(event, stackTrace: stackTrace, hint: hint));
         // we don't call Zone.current.handleUncaughtError because we'd like
         // to set a specific mechanism for FlutterError.onError.
       } else {